# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-harmonicmean",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the harmonic mean of `values` as an `f64`.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. reject the empty list with `422` (no dependency calls);\n2. fold `srvcs-floatadd` over `reciprocal(v)` for each element, starting at\n   `0`, where each reciprocal comes from `srvcs-reciprocal`;\n3. ask `srvcs-floatdivide` for `n / sumrecip` — the harmonic mean.\n\n`n` (the element count) is a trivial local index. If a dependency is\nunreachable it reports itself degraded (`503`); if a dependency rejects the\ninput it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HarmonicMeanResponse"
                }
              }
            }
          },
          "422": {
            "description": "the list is empty, or a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The list of numbers to take the harmonic mean of. Must be non-empty."
          }
        }
      },
      "HarmonicMeanResponse": {
        "type": "object",
        "required": [
          "values",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double"
          },
          "values": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}